6e: split the history persistence pipeline out of HistoryService
HistoryService was a headless JSON pipeline tangled up with a WPF service. The
pipeline half is now Text-Grab.Core.Windows/Utilities/HistoryFileUtilities.cs -
loading and writing the two history files, item-by-item recovery when one is
corrupt, the LanguageKind converter and its AsyncLocal rewrite flag, the
word-border sidecar files, the two normalization passes, retention, and artifact
deletion. All static, no state past the serializer options.

HistoryService keeps its name (every call site is Singleton<HistoryService>) and
keeps what actually held it in the app: the in-memory lists, the two
DispatcherTimers that debounce writes and release the idle cache, the cached
fullscreen bitmap, the recent-grabs MenuItem building, and the SaveToHistory
overloads that take a GrabFrame and an EditTextWindow.

Two behaviour-preserving details worth naming:

- NormalizeHistoryIds used to call MarkHistoryDirty itself; it now returns a bool
  so it can be static. Its callers evaluate it and NormalizeHistoryCompatibilityData
  into locals first - both normalizers mutate, so neither may be short-circuited
  away by the other, which || would have done.
- GetWordBorderInfosAsync stays on the service as a wrapper, because the
  TouchHistoryCache() it starts with is cache bookkeeping, not file work.

PersistWordBorderData reads EnableFileBackedManagedSettings, so that joins
ITextGrabSettings. GetMostRecentGrab and GetExcessVisualHistoryItems moved with
the pipeline, so HistoryServiceTests calls them on the new type.

Also strips BOMs this session's tooling added to five files that never had one.
