+++ Tests/GrabFrameFileTests.cs
@@ -1,4 +1,4 @@
-﻿using System.Drawing;
+using System.Drawing;
 using System.IO;
 using System.IO.Compression;
 using System.Text.Json;
