+++ Text-Grab/Views/EditTextWindow.xaml.cs
@@ -1,4 +1,4 @@
-﻿using Humanizer;
+using Humanizer;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
